gadgets/trace_exec: Add wasm module to handle args field
The args field is encoded by the eBPF program as:
arg0\0arg1\0arg2, this commit implements the logic
to decode and convert it to a string using wasm.

TODO: The datasource doesn't support arrays yet, hence we have to
join the args in a single string. This could be wrong as it's
possible to execute a process with arguments that contain spaces.

TODO2: Are the args size and args count fields really needed? Can't we
imply them?, like two consecutive null bytes means the string ends, etc.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>